Dedupe wallet types
