+++ app/src/os/services/ship/wallet/protocols/BaseBlockProtocol.ts
@@ -1,5 +1,4 @@
-import { Asset } from 'renderer/stores/models/wallet.model';
-
+import { Asset } from '../wallet.types';
 import { BaseProtocol } from './BaseProtocol';
 
 /**
