+++ app/src/renderer/apps/Wallet/WalletApp.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useMemo, useState } from 'react';
 import { observer } from 'mobx-react';
 
-import { Flex } from '@holium/design-system';
+import { Flex } from '@holium/design-system/general';
 
+import { NetworkType } from 'os/services/ship/wallet/wallet.types';
 import {
   BitcoinWalletType,
   EthWalletType,
-  NetworkType,
   TransactionType,
   WalletView,
 } from 'renderer/stores/models/wallet.model';
@@ -123,7 +123,7 @@ const WalletAppPresenter = () => {
 
   return (
     <Flex
-      onClick={(evt: any) => evt.stopPropagation()}
+      onClick={(evt) => evt.stopPropagation()}
       position="relative"
       height={dimensions.height - 24}
       width="100%"
@@ -136,12 +136,8 @@ const WalletAppPresenter = () => {
         network={
           walletStore.navState.network === 'ethereum' ? 'ethereum' : 'bitcoin'
         }
-        onAddWallet={async () =>
-          await walletStore.navigate(WalletView.CREATE_WALLET)
-        }
-        onSetNetwork={async (network: any) =>
-          await walletStore.setNetwork(network)
-        }
+        onAddWallet={() => walletStore.navigate(WalletView.CREATE_WALLET)}
+        onSetNetwork={(network) => walletStore.setNetwork(network)}
         hide={hideHeader}
       />
       {pendingIsVisible &&
