+++ app/src/renderer/stores/ship.store.ts
@@ -2,6 +2,12 @@ import { createContext, useContext } from 'react';
 import { flow, Instance, onSnapshot, SnapshotIn, types } from 'mobx-state-tree';
 
 import { RealmSessionCredentials } from 'os/realm.types';
+import {
+  NetworkStoreType,
+  ProtocolType,
+  SharingMode,
+  WalletCreationMode,
+} from 'os/services/ship/wallet/wallet.types';
 
 import { ChatStore } from './chat.store';
 import { ShipIPC } from './ipc';
@@ -12,14 +18,7 @@ import { FeaturedStore } from './models/featured.model';
 import { FriendsStore } from './models/friends.model';
 import { NotifStore } from './models/notification.model';
 import { SpacesStore } from './models/spaces.model';
-import {
-  NetworkStoreType,
-  ProtocolType,
-  SharingMode,
-  WalletCreationMode,
-  WalletStore,
-  WalletView,
-} from './models/wallet.model';
+import { WalletStore, WalletView } from './models/wallet.model';
 import { RoomsStore } from './rooms.store';
 
 export const ShipStore = types
